xilpm: versal_common: server: Fix MISRA C-2012 Rule 10.4
This patch fixes the MISRA C-2012 Rule 10.4.

Problem: Essential type of the left hand operand is not the same as
that of the right operand.

Solution: Convert to unsigned integer.

Acked-by: Madhav Bhatt <madhav.bhatt@xilinx.com>

+++ lib/sw_services/xilpm/src/versal_common/server/xpm_api.c
@@ -3210,7 +3210,7 @@ XStatus XPm_AbortSuspend(const u32 SubsystemId, const u32 Reason,
 		goto done;
 	}
 
-	if ((Reason < ABORT_REASON_MIN) || (Reason > ABORT_REASON_MAX)) {
+	if ((Reason < (u32)ABORT_REASON_MIN) || (Reason > (u32)ABORT_REASON_MAX)) {
 		Status = XST_INVALID_PARAM;
 		goto done;
 	}
